backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0
python-multipart==0.0.6
openai==1.3.5
pinecone-client==2.2.4
google-api-python-client==2.100.0
google-auth-httplib2==0.1.1
PyGithub==1.59.1
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9
aiofiles==23.2.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
jinja2==3.1.2
markdown==3.5.1
weasyprint==60.2
pandas==2.1.3
numpy==1.26.2
matplotlib==3.8.2
seaborn==0.13.0
scipy==1.11.4
scikit-learn==1.3.2
jupyter==1.0.0
nbconvert==7.11.0
nbformat==5.9.2
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
httpx==0.25.2
python-dotenv==1.0.0
asyncpg
aiosqlite